+++ euphonic_sqw_models/euphonic_wrapper.py
@@ -17,21 +17,14 @@ class CoherentCrystal(object):
     """
     A class to calculate neutron scattering dispersions surfaces for
     use in Horace. It must be constructed from a Euphonic
-    ForceConstants object.
-
-    Methods
-    -------
-    w, sf = horace_disp(qh: np.ndarray, qk: np.ndarray, ql: np.ndarray,
-                        intensity_scale: float = 1.0,
-                        frequency_scale: float = 1.0)
-        Calculates the phonon dispersion surfaces `w` and structure
-        factor `sf` at the specified (qh, qk, ql) points, with optional
-        intensity and frequency scaling factors
+    ForceConstants or euphonic.brille.BrilleInterpolator object.
 
     Attributes
     ----------
-    force_constants : ForceConstants
-        The force constants
+    force_constants : Union[ForceConstants, BrilleInterpolator]
+        An object that has a calculate_qpoint_phonon_modes function
+        that produces a euphonic.QpointPhononModes object. Typically
+        ForceConstants or BrilleInterpolator.
     debye_waller_grid : (3,) float ndarray or None (default: None)
         The Monkhorst-Pack q-point grid size to use for the
         Debye-Waller calculation
@@ -68,7 +61,7 @@ class CoherentCrystal(object):
         database (internal or from a file)
     verbose : bool (default: True)
         Whether to print information on calculation progress
-    **calc_modes_kwargs
+    calc_modes_kwargs
         Any other keyword arguments (e.g asr, n_threads) will be
         passed to euphonic.ForceConstants.calculate_qpoint_phonon_modes.
         See the available arguments with
@@ -76,7 +69,7 @@ class CoherentCrystal(object):
     """
     def __init__(
             self,
-            force_constants: ForceConstants,
+            force_constants: Union[ForceConstants, 'BrilleInterpolator'],
             debye_waller_grid: Optional[Tuple[int, int, int]] = None,
             debye_waller: Optional[DebyeWaller] = None,
             temperature: Union[float, Quantity, None] = None,
@@ -88,13 +81,14 @@ class CoherentCrystal(object):
             scattering_lengths: Union[
                 str, Dict[str, Union[Quantity, float]]] = 'Sears1992',
             verbose: bool = True,
-            n_threads: Optional[int] = None,
             **calc_modes_kwargs) -> None:
         """
         Parameters
         ----------
         force_constants
-            The force constants
+            An object that has a calculate_qpoint_phonon_modes function
+            that produces a euphonic.QpointPhononModes object.
+            Typically ForceConstants or BrilleInterpolator.
         debye_waller_grid
             The Monkhorst-Pack q-point grid size to use for the
             Debye-Waller calculation. If None the Debye-Waller
@@ -132,13 +126,10 @@ class CoherentCrystal(object):
             string denoting a database (internal or from a file)
         verbose
             Whether to print information on calculation progress
-        n_threads
-            The number of threads to use in the interpolation
         **calc_modes_kwargs
             Any other keyword arguments (e.g asr, use_c) will be
-            passed to euphonic.ForceConstants.calculate_qpoint_phonon_modes.
-            See the available arguments with
-            help(euphonic.ForceConstants.calculate_qpoint_phonon_modes)
+            passed to the calculate_qpoint_phonon_modes method of
+            the input ForceConstants or BrilleInterpolator object.
         """
         self.force_constants = force_constants
         self.debye_waller_grid = debye_waller_grid
@@ -154,10 +145,12 @@ class CoherentCrystal(object):
         self.lim = lim
         self.scattering_lengths = scattering_lengths
         self.verbose = verbose
-        # Explicitly have n_threads as an argument so we can use typing
-        # to automatically convert to int when called from MATLAB
-        if n_threads is not None:
-            calc_modes_kwargs['n_threads'] = n_threads
+        # Explicitly convert to integers so users calling from MATLAB
+        # don't have to
+        int_kwargs = ['threads', 'n_threads']
+        for ikw in int_kwargs:
+            if ikw in calc_modes_kwargs:
+                calc_modes_kwargs[ikw] = int(calc_modes_kwargs[ikw])
         self.calc_modes_kwargs = calc_modes_kwargs
 
     def _calculate_sf(self, qpts: np.ndarray
@@ -224,9 +217,10 @@ class CoherentCrystal(object):
                 if self.verbose:
                     print(f'Using Euphonic to interpolate for q-points '
                           f'{qi}:{qf} out of {lqh}')
-                qpts = np.vstack((np.squeeze(qh[qi:qf]),
-                                  np.squeeze(qk[qi:qf]),
-                                  np.squeeze(ql[qi:qf]))).T
+                qpts = np.stack((np.squeeze(qh[qi:qf]),
+                                 np.squeeze(qk[qi:qf]),
+                                 np.squeeze(ql[qi:qf])),
+                                 axis=1)
                 if self.conversion_mat is not None:
                     qpts = np.matmul(qpts, self.conversion_mat)
                 sqw = self._calculate_sf(qpts)
@@ -237,9 +231,10 @@ class CoherentCrystal(object):
                     w = np.vstack((w, sqw[0]))
                     sf = np.vstack((sf, sqw[1]))
         else:
-            qpts = np.vstack((np.squeeze(qh),
-                              np.squeeze(qk),
-                              np.squeeze(ql))).T
+            qpts = np.stack((np.squeeze(qh),
+                             np.squeeze(qk),
+                             np.squeeze(ql)),
+                             axis=1)
             if self.conversion_mat is not None:
                 qpts = np.matmul(qpts, self.conversion_mat)
             w, sf = self._calculate_sf(qpts)
@@ -281,11 +276,12 @@ class CoherentCrystal(object):
         return chunk
 
     @property
-    def force_constants(self) -> ForceConstants:
+    def force_constants(self) -> Union[ForceConstants, 'BrilleInterpolator']:
         return self._force_constants
 
     @force_constants.setter
-    def force_constants(self, val: ForceConstants) -> None:
+    def force_constants(self, val: Union[ForceConstants, 'BrilleInterpolator']
+                        ) -> None:
         if hasattr(val, 'calculate_qpoint_phonon_modes'):
             self._force_constants = val
         else: